Add Jpa Concept

+++ socialmedia/src/main/java/com/example/restfulwebservices/jdbc/Course.java
@@ -7,7 +7,7 @@ import jakarta.persistence.Id;
 public class Course {
 
 	@Id
-	private int id;
+	private long id;
 	private String name;
 	private String author;
 
@@ -15,14 +15,14 @@ public class Course {
 
 	}
 
-	public Course(int id, String name, String author) {
+	public Course(long id, String name, String author) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.author = author;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
